Tests/LibWeb/Text/expected/HTML/Window-close.txt

window.closed = false
window.closed = true
window.closed = true
